refactor(db): extract pure validation from organizations module
Convert db/organizations.rs (3,513 lines) into a directory module and
move the synchronous, side-effect-free validation layer into
organizations/validation.rs: RESERVED_TLDS, DomainValidationError,
normalize_domain, unicode_form, RESERVED_SUBDOMAIN_LABELS,
SubdomainLabelError, validate_subdomain_label, and the verified-apex
label derivation (eligible_subdomain_labels,
ineligible_subdomain_candidates, backing_apex_for_label), together with
their 14 unit tests.

organizations/mod.rs re-exports the public names, so db:: paths and
every caller stay unchanged. Persistence, the claim/release state
machines, and signing-key storage remain in mod.rs; issues #636 and
#637 track those larger seams.
